(PUP-10439) Collapse Pool and BasePool
