+++ lib/puppet/http.rb
@@ -1,7 +1,6 @@
 module Puppet
   module Network
     module HTTP
-      require 'puppet/network/http/base_pool'
       require 'puppet/network/http/pool'
       require 'puppet/network/resolver'
     end
